+++ examples/ganglionServer/ganglionServer.js
@@ -108,6 +108,17 @@ function exitHandler (options, err) {
   }
 }
 
+if (process.platform === "win32") {
+  const rl = require("readline").createInterface({
+    input: process.stdin,
+    output: process.stdout
+  });
+
+  rl.on("SIGINT", function () {
+    process.emit("SIGINT");
+  });
+}
+
 // do something when app is closing
 process.on('exit', exitHandler.bind(null, {
   cleanup: true