+++ test/openBCIGanglion-test.js
@@ -45,24 +45,24 @@ describe('#ganglion', function () {
         0b00000000, // 1
         0b00000000, // 2
         0b00000000, // 3
-        0b00001000, // 4
+        0b00100000, // 4
         0b00000000, // 5
-        0b00000101, // 6
+        0b00101000, // 6
         0b00000000, // 7
-        0b00000000, // 8
-        0b01001000, // 9
+        0b00000100, // 8
+        0b10000000, // 9
         0b00000000, // 10
-        0b00001001, // 11
-        0b11110000, // 12
-        0b00000001, // 13
-        0b10110000, // 14
-        0b00000000, // 15
-        0b00110000, // 16
-        0b00000000, // 17
-        0b00001000  // 18
+        0b10111100, // 11
+        0b00000000, // 12
+        0b00000111, // 13
+        0b00000000, // 14
+        0b00101000, // 15
+        0b11000000, // 16
+        0b00001010  // 17
       ]);
-    let expectedValue = [[0, 2, 10, 4], [262148, 507910, 393222, 8]];
+    let expectedValue = [[0, 2, 10, 4], [131074, 245760, 114698, 49162]];
     let actualValue = ganglion._decompressDeltas(buffer);
+
     for (let i = 0; i < 4; i++) {
       (actualValue[0][i]).should.equal(expectedValue[0][i]);
       (actualValue[1][i]).should.equal(expectedValue[1][i]);
@@ -73,23 +73,22 @@ describe('#ganglion', function () {
       [
         0b11111111, // 0
         0b11111111, // 1
-        0b10111111, // 2
+        0b01111111, // 2
         0b11111111, // 3
-        0b11101111, // 4
+        0b10111111, // 4
         0b11111111, // 5
-        0b11111100, // 6
+        0b11100111, // 6
         0b11111111, // 7
-        0b11111111, // 8
-        0b01011000, // 9
-        0b00000000, // 10
-        0b00001011, // 11
-        0b00111110, // 12
-        0b00111000, // 13
-        0b11100000, // 14
-        0b00000000, // 15
-        0b00111111, // 16
-        0b11110000, // 17
-        0b00000001  // 18
+        0b11110101, // 8
+        0b00000000, // 9
+        0b00000001, // 10
+        0b01001111, // 11
+        0b10001110, // 12
+        0b00110000, // 13
+        0b00000000, // 14
+        0b00011111, // 15
+        0b11110000, // 16
+        0b00000001  // 17
       ]);
     let expectedValue = [[-3, -5, -7, -11], [-262139, -198429, -262137, -4095]];
     let actualValue = ganglion._decompressDeltas(buffer);
@@ -181,41 +180,124 @@ describe('#ganglion', function () {
       ganglion.removeListener('accelerometer', accelDataFunc);
     });
   });
+  describe('#_processProcessSampleData', function () {
+    let funcSpyCompressedData;
+    let funcSpyDroppedPacket;
+    let funcSpyUncompressedData;
+    before(function () {
+      funcSpyCompressedData = sinon.spy(ganglion, '_processCompressedData');
+      funcSpyDroppedPacket = sinon.spy(ganglion, '_droppedPacket');
+      funcSpyUncompressedData = sinon.spy(ganglion, '_processUncompressedData');
+    });
+    beforeEach(function () {
+      funcSpyCompressedData.reset();
+      funcSpyDroppedPacket.reset();
+      funcSpyUncompressedData.reset();
+      ganglion._resetDroppedPacketSystem();
+    });
+    it('should call proper functions if no dropped packets', function () {
+      it('should work on uncompressed data', function () {
+        ganglion._processProcessSampleData(utils.sampleUncompressedData());
+        funcSpyUncompressedData.should.have.been.called;
+        funcSpyDroppedPacket.should.not.have.been.called;
+      });
+
+      it('should work on compressed data', function () {
+        ganglion._processProcessSampleData(utils.sampleCompressedData(1));
+        funcSpyCompressedData.should.have.been.called;
+        funcSpyDroppedPacket.should.not.have.been.called;
+      });
+    });
+    it('should recognize 0 packet dropped', function () {
+      // Send the last buffer, set's ganglion._packetCounter
+      ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax));
+      funcSpyCompressedData.should.have.been.called;
+      const expectedMissedSample = k.OBCIGanglionByteIdUncompressed;
+      // Call the function under test with one more then expected
+      const nextPacket = utils.sampleCompressedData(expectedMissedSample + 1);
+      ganglion._processProcessSampleData(nextPacket);
+      funcSpyCompressedData.should.have.been.calledTwice;
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample);
+    });
+    it('should not find a dropped packet on wrap around', function () {
+      ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax - 1));
+      funcSpyCompressedData.should.have.been.calledOnce;
+      ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax));
+      funcSpyCompressedData.should.have.been.calledTwice;
+      ganglion._processProcessSampleData(utils.sampleUncompressedData());
+      funcSpyCompressedData.should.have.been.calledTwice;
+      funcSpyUncompressedData.should.have.been.calledOnce;
+      ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdUncompressed + 1));
+      funcSpyCompressedData.should.have.been.calledThrice;
+      funcSpyDroppedPacket.should.not.have.been.called;
+    });
+    it('should try to resend packet 127', function () {
+      ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax - 1));
+      const expectedMissedSample = k.OBCIGanglionByteIdSampleMax;
+      // Call the function under test with one more then expected
+      const nextPacket = utils.sampleUncompressedData();
+      ganglion._processProcessSampleData(nextPacket);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample);
+    });
+    it('should try to resend packet 126 and 127', function () {
+      ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax - 2));
+      const expectedMissedSample1 = k.OBCIGanglionByteIdSampleMax - 1;
+      const expectedMissedSample2 = k.OBCIGanglionByteIdSampleMax;
+      // Call the function under test with one more then expected
+      const nextPacket = utils.sampleUncompressedData();
+      ganglion._processProcessSampleData(nextPacket);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample1);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample2);
+    });
+    it('should try to resend packet 1 and add packet 2 to buffer', function () {
+      // Send the last buffer, set's ganglion._packetCounter
+      ganglion._processProcessSampleData(utils.sampleUncompressedData());
+      const expectedMissedSample = k.OBCIGanglionByteIdUncompressed + 1;
+      // Call the function under test with one more then expected
+      const nextPacket = utils.sampleCompressedData(expectedMissedSample + 1);
+      ganglion._processProcessSampleData(nextPacket);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample);
+    });
+    it('should try to resend packet 1 & 2 and add packet 3 to buffer', function () {
+      // Send the last buffer, set's ganglion._packetCounter
+      ganglion._processProcessSampleData(utils.sampleUncompressedData());
+      const expectedMissedSample1 = k.OBCIGanglionByteIdUncompressed + 1;
+      const expectedMissedSample2 = k.OBCIGanglionByteIdUncompressed + 2;
+      // Call the function under test with two more then expected
+      const nextPacket = utils.sampleCompressedData(expectedMissedSample2 + 1);
+      ganglion._processProcessSampleData(nextPacket);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample1);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample2);
+    });
+  });
   describe('_processBytes', function () {
     let funcSpyAccel;
-    let funcSpyCompressedData;
     let funcSpyImpedanceData;
     let funcSpyMultiBytePacket;
     let funcSpyMultiBytePacketStop;
     let funcSpyOtherData;
-    let funcSpyUncompressedData;
+    let funcSpyProcessedData;
 
     before(function () {
       // Put watchers on all functions
       funcSpyAccel = sinon.spy(ganglion, '_processAccel');
-      funcSpyCompressedData = sinon.spy(ganglion, '_processCompressedData');
       funcSpyImpedanceData = sinon.spy(ganglion, '_processImpedanceData');
       funcSpyMultiBytePacket = sinon.spy(ganglion, '_processMultiBytePacket');
       funcSpyMultiBytePacketStop = sinon.spy(ganglion, '_processMultiBytePacketStop');
       funcSpyOtherData = sinon.spy(ganglion, '_processOtherData');
-      funcSpyUncompressedData = sinon.spy(ganglion, '_processUncompressedData');
+      funcSpyProcessedData = sinon.spy(ganglion, '_processProcessSampleData');
     });
     beforeEach(function () {
       funcSpyAccel.reset();
-      funcSpyCompressedData.reset();
       funcSpyImpedanceData.reset();
       funcSpyMultiBytePacket.reset();
       funcSpyMultiBytePacketStop.reset();
       funcSpyOtherData.reset();
-      funcSpyUncompressedData.reset();
+      funcSpyProcessedData.reset();
     });
     it('should route accel packet', function () {
       ganglion._processBytes(utils.sampleAccel());
       funcSpyAccel.should.have.been.calledOnce;
-    });
-    it('should route compressed data packet', function () {
-      ganglion._processBytes(utils.sampleCompressedData(3));
-      funcSpyCompressedData.should.have.been.calledOnce;
     });
     it('should route impedance channel 1 packet', function () {
       ganglion._processBytes(utils.sampleImpedanceChannel1());
@@ -249,9 +331,9 @@ describe('#ganglion', function () {
       ganglion._processBytes(utils.sampleOtherData(new Buffer('blah')));
       funcSpyOtherData.should.have.been.calledOnce;
     });
-    it('should route uncompressed data packet', function () {
+    it('should route processed data packet', function () {
       ganglion._processBytes(utils.sampleUncompressedData());
-      funcSpyUncompressedData.should.have.been.calledOnce;
+      funcSpyProcessedData.should.have.been.calledOnce;
     });
   });
   it('should emit impedance value', function () {
@@ -338,6 +420,7 @@ describe('#ganglion', function () {
     // Makes sure the correct amount of events were called.
     expect(runningEventCount).to.equal(totalEvents);
   });
+
 });
 
 xdescribe('#noble', function () {