+++ changelog.md
@@ -9,6 +9,7 @@
 ### Bug Fixes
 * Fix bug where ganglionServer example would hang on scan if no board found.
 * Fix bug where node process would not disconnect on windows.
+* Fix unhandled promise where server would call search start twice.
 
 # 0.1.1